nowhub/requirements.txt
django==2.1.7
djangorestframework==3.9.2
coreapi==2.3.3
psycopg2-binary==2.7.7
google-cloud-firestore==0.31.0
